ints are fine

+++ 031/c/solution.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
-long solve(int p);
-long go (int amount, int first);
+int solve(int p);
+int go (int amount, int first);
 
 #ifndef TESTING
 int main () {
-    printf("%ld\n", solve(200));
+    printf("%d\n", solve(200));
     return 0;
 }
 #endif
@@ -13,11 +13,11 @@ int main () {
 #define NUMCOINS 8
 static int coins[NUMCOINS] = {1, 2, 5, 10, 20, 50, 100, 200};
 
-long solve(int p) {
+int solve(int p) {
   return go(p, 0);
 }
 
-long go(int amount, int first) {
+int go(int amount, int first) {
   if (first == NUMCOINS) { return amount == 0 ? 1 : 0; }
   int cur = coins[first];
   int res = 0, i;
